+++ addon/components/paper-fab-actions.js
@@ -1,0 +1,5 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  tagName: 'md-fab-actions',
+});