+++ addon/components/paper-fab-trigger.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  tagName: 'md-fab-trigger',
+
+  mouseEnter() {
+    this.sendAction('mouse-enter');
+  }
+});