+++ tests/dummy/app/controllers/fab-speed-dial.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+
+export default Ember.Controller.extend({
+  selectedVAlignDirection: 'bottom',
+  selectedHAlignDirection: 'right',
+  selectedDirection: 'up',
+  selectedDirection2: 'up',
+  selectedAnimation: 'fling',
+  selectedOpen: false,
+  selectedHover: false,
+});